Changed item struct datatypes to reflect definitions in TDD

+++ header.h
@@ -3,5 +3,5 @@ typedef struct {
     char description[41];   //Item Description
     unsigned int qty;       //Item Quantity
     char exp[10];           //Item Expiry Date
-    double price;           //Item price
+    float price;           //Item price
   } item;
